move nextjs specific rl check into web
