+++ apps/web/server/lib/[user]/getServerSideProps.ts
@@ -8,7 +8,7 @@ import { getUsernameList } from "@calcom/features/eventtypes/lib/defaultEvents";
 import { getEventTypesPublic } from "@calcom/features/eventtypes/lib/getEventTypesPublic";
 import { getBrandingForUser } from "@calcom/features/profile/lib/getBranding";
 import { UserRepository } from "@calcom/features/users/repositories/UserRepository";
-import { handleRateLimitForNextJs } from "@calcom/lib/checkRateLimitAndThrowError";
+import { handleRateLimitForNextJs } from "@lib/handleRatelimit";
 import { DEFAULT_DARK_BRAND_COLOR, DEFAULT_LIGHT_BRAND_COLOR } from "@calcom/lib/constants";
 import { getUserAvatarUrl } from "@calcom/lib/getAvatarUrl";
 import getIP from "@calcom/lib/getIP";
